(Client) Rename `models` to `fullstack-shared-models`, hook up syncing between projects
