+++ client-v2/package.json
@@ -7,6 +7,7 @@
         "build": "ng build",
         "watch": "ng build --watch --configuration development",
         "update-colors": "ts-node ./tools/update-colors.cts",
+        "sync-shared": "rsync -avz src/app/fullstack-shared-models ../server/src/",
         "lint": "eslint \"{src,app,test,cypress}/**/**.{ts,html}\"",
         "lint:fix": "npm run lint -- --fix",
         "unit": "ng test",
