+++ client-v2/src/app/components/organisms/entity-view/entity-view.component.test.ts
@@ -1,8 +1,8 @@
 import { CdkMenuModule } from '@angular/cdk/menu'
 import { Injector } from '@angular/core'
 import { BehaviorSubject } from 'rxjs'
-import { EntityPreviewRecursive } from 'src/app/models/entities.model'
-import { TasklistDetail } from 'src/app/models/list.model'
+import { EntityPreviewRecursive } from 'src/app/fullstack-shared-models/entities.model'
+import { TasklistDetail } from 'src/app/fullstack-shared-models/list.model'
 import { storeMock } from 'src/app/utils/unit-test.mocks'
 import { EntityPageLabelComponent } from '../../atoms/entity-page-label/entity-page-label.component'
 import { DropDownComponent } from '../../molecules/drop-down/drop-down.component'
