+++ client-v2/src/app/fullstack-shared-models/list.model.ts
