+++ client-v2/src/app/fullstack-shared-models/task.model.ts
@@ -1,3 +1,5 @@
+import { EntityPreview, EntityPreviewRecursive, EntityType } from './entities.model'
+
 // @TODO: Update these fields in the backend
 export enum TaskStatus {
     BACKLOG = 'Backlog',
