+++ client-v2/src/app/pages/auth/login/login.component.ts
@@ -5,7 +5,7 @@ import { Store } from '@ngrx/store'
 import { map } from 'rxjs'
 import { FormBuilderOptions } from 'src/app/components/molecules/form/types'
 import { betterEmailValidator } from 'src/app/components/molecules/form/validators'
-import { LoginCredentialsDto } from 'src/app/models/auth.model'
+import { LoginCredentialsDto } from 'src/app/fullstack-shared-models/auth.model'
 import { AppState } from 'src/app/store'
 import { authActions } from 'src/app/store/user/user.actions'
 import { userSelectors } from 'src/app/store/user/user.selectors'
