+++ client-v2/src/app/pages/home/entity-page/entity-page.component.ts
@@ -5,7 +5,7 @@ import { Store } from '@ngrx/store'
 import { map, of, shareReplay, switchMap } from 'rxjs'
 import { Breadcrumb } from 'src/app/components/molecules/breadcrumbs/breadcrumbs.component'
 import { useDataForAction } from 'src/app/components/molecules/drop-down/drop-down.component'
-import { EntityType } from 'src/app/models/entities.model'
+import { EntityType } from 'src/app/fullstack-shared-models/entities.model'
 import { getEntityMenuItemsMap } from 'src/app/shared/entity-menu-items'
 import { AppState } from 'src/app/store'
 import { traceEntity } from 'src/app/store/entities/utils'
