+++ client-v2/src/app/pages/home/home.component.ts
@@ -5,7 +5,7 @@ import { ActivatedRoute } from '@angular/router'
 import { Actions } from '@ngrx/effects'
 import { Store } from '@ngrx/store'
 import { map, startWith, switchMap, tap } from 'rxjs'
-import { EntityPreviewFlattend, EntityType } from 'src/app/models/entities.model'
+import { EntityPreviewFlattend, EntityType } from 'src/app/fullstack-shared-models/entities.model'
 import { getEntityMenuItemsMap } from 'src/app/shared/entity-menu-items'
 import { AppState } from 'src/app/store'
 import { entitiesActions, listActions } from 'src/app/store/entities/entities.actions'
