+++ client-v2/src/app/shared/defaults.ts
@@ -1,5 +1,6 @@
-import { EntityType } from './entities.model'
+import { EntityType } from '../fullstack-shared-models/entities.model'
 
 export const ENTITY_TITLE_DEFAULTS: Record<EntityType, string> = {
     [EntityType.TASKLIST]: 'Untitled tasklist',
+    // [EntityType.TASK]: 'Untitled task',
 }
