+++ client-v2/src/app/shared/models/index.ts
@@ -1,0 +1,4 @@
+import { LoggedInUser } from 'src/app/fullstack-shared-models/auth.model'
+import { HttpSuccessResponse } from 'src/app/http/types'
+
+export type AuthSuccessResponse = HttpSuccessResponse<{ user: LoggedInUser }>
