+++ client-v2/src/app/store/entities/entities.effects.ts
@@ -5,7 +5,7 @@ import { Actions, createEffect, ofType } from '@ngrx/effects'
 import { Store } from '@ngrx/store'
 import { catchError, concatMap, first, map, mergeMap, of, switchMap, tap } from 'rxjs'
 import { DialogService } from 'src/app/modal/dialog.service'
-import { EntityType } from 'src/app/models/entities.model'
+import { EntityType } from 'src/app/fullstack-shared-models/entities.model'
 import { EntitiesService } from 'src/app/services/entities.service'
 import { getMessageFromHttpError } from 'src/app/utils/store.helpers'
 import { AppState } from '..'
