+++ client-v2/src/app/store/entities/list.effects.ts
@@ -5,8 +5,8 @@ import { Actions, createEffect, ofType } from '@ngrx/effects'
 import { Store } from '@ngrx/store'
 import { catchError, map, mergeMap, of, tap } from 'rxjs'
 import { DialogService } from 'src/app/modal/dialog.service'
-import { ENTITY_TITLE_DEFAULTS } from 'src/app/models/defaults'
-import { EntityType } from 'src/app/models/entities.model'
+import { ENTITY_TITLE_DEFAULTS } from 'src/app/shared/defaults'
+import { EntityType } from 'src/app/fullstack-shared-models/entities.model'
 import { ListService } from 'src/app/services/entity.services/list.service'
 import { getMessageFromHttpError } from 'src/app/utils/store.helpers'
 import { AppState } from '..'
