+++ client-v2/src/app/store/entities/task.effects.ts
@@ -4,8 +4,8 @@ import { Actions, createEffect, ofType } from '@ngrx/effects'
 import { Store } from '@ngrx/store'
 import { catchError, map, mergeMap, of } from 'rxjs'
 import { DialogService } from 'src/app/modal/dialog.service'
-import { ENTITY_TITLE_DEFAULTS } from 'src/app/models/defaults'
-import { EntityType } from 'src/app/models/entities.model'
+import { ENTITY_TITLE_DEFAULTS } from 'src/app/shared/defaults'
+import { EntityType } from 'src/app/fullstack-shared-models/entities.model'
 import { TaskService } from 'src/app/services/entity.services/task.service'
 import { getMessageFromHttpError } from 'src/app/utils/store.helpers'
 import { AppState } from '..'
