+++ server/package.json
@@ -9,6 +9,7 @@
         "prebuild": "rimraf dist",
         "build": "nest build",
         "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
+        "sync-shared": "cd ../client-v2 && npm run sync-shared",
         "prestart": "npx prisma migrate deploy",
         "start": "nest start",
         "start:testing": "PORT=3001 TESTING_ENV=true nest start",
